tidy: move directory traversal utility functions into dedicated module
